requirements.txt
django
psycopg2
django-bootstrap3
djangorestframework